getFourExercises func
